+++ test/adapters/document-sync-adapter.test.ts
@@ -1,5 +1,6 @@
 import { TextDocumentSyncKind, TextDocumentSyncOptions } from "../../lib/languageclient"
 import DocumentSyncAdapter from "../../lib/adapters/document-sync-adapter"
+import { createFakeEditor } from "../helpers"
 
 describe("DocumentSyncAdapter", () => {
   describe("canAdapt", () => {
@@ -52,7 +53,8 @@ describe("DocumentSyncAdapter", () => {
         null as any,
         () => false,
         textDocumentSync,
-        (_t, f) => f()
+        (_t, f) => f(),
+        () => ""
       )
     }
 
@@ -79,6 +81,26 @@ describe("DocumentSyncAdapter", () => {
     it("sets _documentSync.change correctly Full for unset capabilities", () => {
       const result = create()._documentSync.change
       expect(result).toBe(TextDocumentSyncKind.Full)
+    })
+  })
+
+  describe("getLanguageId", () => {
+    function create(getLanguageIdFromEditor: () => string) {
+      return new DocumentSyncAdapter(
+        null as any,
+        () => false,
+        TextDocumentSyncKind.Incremental,
+        (_t, f) => f(),
+        getLanguageIdFromEditor
+      )
+    }
+
+    it("use the return value of `_getLanguageIdFromEditor` as the languageId", () => {
+      const editor = createFakeEditor()
+      const adapter = create(() => "someLanguageId") as any
+      adapter._handleNewEditor(editor)
+      const result = adapter.getEditorSyncAdapter(editor).getLanguageId()
+      expect(result).toBe("someLanguageId")
     })
   })
 })